Updated to timer

+++ Stopwatch.h
@@ -5,18 +5,18 @@ class Stopwatch
 {
     public:
         Stopwatch();
-        void start(int length);
-        int GetElapsed();
+        void start(unsigned long length);
+        unsigned long GetElapsed();
         bool IsRunning();
 
-        int ONE_SECOND = 1 * 1000;
-        int ONE_MINUTE = 1 * 60 * ONE_SECOND;
-        int ONE_HOUR = 1 * 60 * ONE_MINUTE;
+        unsigned long ONE_SECOND = 1 * 1000;
+        unsigned long ONE_MINUTE = 1 * 60 * ONE_SECOND;
+        unsigned long ONE_HOUR = 1 * 60 * ONE_MINUTE;
 
     private:
         bool isRunning = false;
         unsigned long StartTime;
-        int threshold;
+        unsigned long threshold;
 };
